github: combine token + ref services into GitHubProjectRepoService
Both services were @Publish'd through the same auth path with identical
project + install resolution. Splitting them wasn't actually buying any
capability separation — it just duplicated the (orgId, projectId) ->
(Project, GitHubAppInstallation) lookup. Folded into one service with
issueRepoToken, createTag, createBranch.
