+++ kinotic-github/src/main/java/org/kinotic/github/api/services/GitHubProjectRepoService.java
@@ -1,0 +1,55 @@
+package org.kinotic.github.api.services;
+
+import org.kinotic.core.api.annotations.Publish;
+import org.kinotic.core.api.annotations.Scope;
+import org.kinotic.github.api.model.GitHubInstallationToken;
+
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * Operations against a Kinotic Project's backing GitHub repository: minting
+ * short-lived installation tokens for worker clones, and creating tags or
+ * branches as part of release flows.
+ * <p>
+ * Scoped by {@code organizationId} so all calls for one org pin to the same node
+ * (keeps the in-memory installation-token cache hot per org).
+ */
+@Publish
+public interface GitHubProjectRepoService {
+
+    /**
+     * Issues a {@code contents:read}-scoped installation token a worker can use
+     * to clone the project's backing repo. The token is short-lived (~1 hour);
+     * workers should not cache it across job runs.
+     *
+     * @param organizationId the caller's Kinotic org id; must equal the authenticated
+     *                       participant's org
+     * @param projectId      the project whose repo the worker wants to clone
+     * @return token + expiry + clone URL, or a failed future if the project has no
+     *         GitHub repo provisioned
+     */
+    CompletableFuture<GitHubInstallationToken> issueRepoToken(@Scope String organizationId,
+                                                              String projectId);
+
+    /**
+     * Creates a lightweight tag on the project's backing repo.
+     *
+     * @param organizationId the caller's Kinotic org id (must match the authenticated
+     *                       participant's org)
+     * @param projectId      the project whose repo to tag
+     * @param tagName        e.g. {@code v1.2.0}
+     * @param sha            full 40-character commit SHA the tag should point at
+     */
+    CompletableFuture<Void> createTag(@Scope String organizationId,
+                                      String projectId,
+                                      String tagName,
+                                      String sha);
+
+    /**
+     * Creates a branch on the project's backing repo pointing at {@code sha}.
+     */
+    CompletableFuture<Void> createBranch(@Scope String organizationId,
+                                         String projectId,
+                                         String branchName,
+                                         String sha);
+}
